+++ docs/src/Link.doc.js
@@ -74,24 +74,6 @@ card(
 card(
   <Example
     description={`
-    \`href\` is optional on \`Link\` so you can use a custom \`Link\` (e.g. for \`react-router\` support).
-  `}
-    name="Example: optional href"
-    defaultCode={`
-<Link>
-  <Box padding={2}>
-    <a href="https://pinterest.com" style={{ textDecoration: 'none' }}>
-      <Text bold>Pinterest.com</Text>
-    </a>
-  </Box>
-</Link>
-`}
-  />
-);
-
-card(
-  <Example
-    description={`
     When providing the content for the link, avoid phrases like "click here" or "go to".
   `}
     name="Accessibility"